Added logging and toggles for database and local testing.
