+++ code/rock_paper_scissors.py
@@ -1,9 +1,12 @@
 import random
 #from code import mongodb_connection as db
-from code import mysql_connection as db
+#from code import mysql_connection as db
 from code.fb_messenger import *
 from code import bot_behaviour
 from time import sleep
+import logging
+import os
+log = logging.getLogger(os.path.basename(__file__))
 
 def play_a_round(userid, user_choice):
     """Function that plays a round of rock - paper - scissors.
@@ -11,7 +14,7 @@ def play_a_round(userid, user_choice):
     winner bot -> 0; winner player -> 1; draw =-> None"""
     choices = ['rock', 'paper', 'scissors']
     bot_choice = random.choice(choices)
-    print("[LOG-GAME-INFO] User chose: " + user_choice + " Bot chose: " + bot_choice)
+    log.info("[LOG-GAME-INFO] User chose: " + user_choice + " Bot chose: " + bot_choice)
     game_outcome = -1
     # establish game outcome
     if user_choice == bot_choice:
@@ -37,7 +40,7 @@ def play_a_round(userid, user_choice):
 
     # establish return statements and update the database
     # winner bot -> 0; winner player -> 1; draw =-> -1"
-    db.update_player_results(userid, game_outcome)
+    #db.update_player_results(userid, game_outcome)
     return [game_outcome, bot_choice]
 
 def play(user_message = "", userid="", bot=""):
@@ -52,10 +55,10 @@ def play(user_message = "", userid="", bot=""):
     if choice == "new_game":
         entry_message = random.choice(["Which one do you choose?", "so, rock, paper or scissors?", "ok, let's play!"])
         bot.fb_send_quick_replies(userid, entry_message, ["✊ rock","✋ paper","✌ scissors"])
-        print("[LOG-GAME-INFO] User #{0} started a new RPS game with: '{1}'.".format(str(userid)[0:4], str(user_message)))
+        log.info("[LOG-GAME-INFO] User #{0} started a new RPS game with: '{1}'.".format(str(userid)[0:4], str(user_message)))
         return "Game started"
     elif choice == "rock" or choice == "paper" or choice == "scissors":
-        print("[LOG-GAME-INFO] User has choosen: "+str(choice))
+        log.info("[LOG-GAME-INFO] User has choosen: "+str(choice))
         game_outcome = play_a_round(userid, choice)
         if game_outcome[0] == -1:
             response = random.choice(["Uff! It's a draw!", "Tie!"])
@@ -66,10 +69,10 @@ def play(user_message = "", userid="", bot=""):
         else:
             response = "I can't play this game."
         bot.fb_send_text_message(userid, game_outcome[1])
-        db.add_conversation(userid, 'Bot', game_outcome[1])
+        #db.add_conversation(userid, 'Bot', game_outcome[1])
         sleep(0.2)
         bot.fb_send_text_message(userid, str(response))
-        db.add_conversation(userid, 'Bot', str(response))
+        #db.add_conversation(userid, 'Bot', str(response))
         return "Game played"
     else:
         return "Error?"
